Moved models to Domain
