+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/EventDetails.razor
@@ -1,5 +1,5 @@
 @page "/events/details/{EventId:int}"
-@using CleanArchitectureBlazor.Models
+@using Domain
 @using CleanArchitectureBlazor.Services
 @inject IEventService EventService
 @inject IShiftService ShiftService
