+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/EventEdit.razor
@@ -1,5 +1,5 @@
 @page "/events/edit/{EventId:int}"
-@using CleanArchitectureBlazor.Models
+@using Domain
 @using CleanArchitectureBlazor.Services
 @inject IEventService EventService
 @inject IShiftService ShiftService
