+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Events.razor
@@ -1,5 +1,5 @@
 @page "/events"
-@using CleanArchitectureBlazor.Models
+@using Domain
 @using CleanArchitectureBlazor.Services
 @inject IEventService EventService
 @inject IShiftService ShiftService
