+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/ShiftDetails.razor
@@ -1,5 +1,5 @@
 @page "/shifts/details/{ShiftId:int}"
-@using CleanArchitectureBlazor.Models
+@using Domain
 @using CleanArchitectureBlazor.Services
 @inject IShiftService ShiftService
 @inject IStaffAssignmentService StaffAssignmentService
