+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/Shifts.razor
@@ -1,5 +1,5 @@
 @page "/shifts"
-@using CleanArchitectureBlazor.Models
+@using Domain
 @using CleanArchitectureBlazor.Services
 @inject IShiftService ShiftService
 @inject IEventService EventService
