+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/Pages/StaffDetails.razor
@@ -1,5 +1,5 @@
 @page "/staff/details/{StaffId:int}"
-@using CleanArchitectureBlazor.Models
+@using Domain
 @using CleanArchitectureBlazor.Services
 @inject IStaffService StaffService
 @inject IStaffAssignmentService StaffAssignmentService
@@ -513,8 +513,8 @@ else
     private bool isDeactivating = false;
     private bool showEditDialog = false;
     private bool showDeactivateConfirm = false;
-    private CleanArchitectureBlazor.Models.Staff? staffDetail;
-    private CleanArchitectureBlazor.Models.Staff staffForm = new();
+    private Domain.Staff? staffDetail;
+    private Domain.Staff staffForm = new();
     private string emailError = string.Empty;
 
     protected override async Task OnInitializedAsync()
@@ -569,7 +569,7 @@ else
     {
         if (staffDetail == null) return;
 
-        staffForm = new CleanArchitectureBlazor.Models.Staff
+        staffForm = new Domain.Staff
         {
             Id = staffDetail.Id,
             FirstName = staffDetail.FirstName,
@@ -591,7 +591,7 @@ else
     private void CloseEditDialog()
     {
         showEditDialog = false;
-        staffForm = new CleanArchitectureBlazor.Models.Staff();
+        staffForm = new Domain.Staff();
         emailError = string.Empty;
         StateHasChanged();
     }
