+++ CleanArchitectureBlazor/CleanArchitectureBlazor/Components/_Imports.razor
@@ -11,3 +11,4 @@
 @using CleanArchitectureBlazor.Client
 @using CleanArchitectureBlazor.Components
 @using CleanArchitectureBlazor.Components.Layout
+@using Domain
